secured backend endpoints such that they ocan only be accessed by authenticated users

+++ backend/api/event.js
@@ -1,7 +1,7 @@
 const express = require('express');
-
 const Event = require('../models/Event');
 const router = express.Router();
+const authEndpoint = require('./auth')
 
 const createEvent = async (eventName, eventDate, eventLocation, eventVolunteerPreferences) => {
     Event.findOne({name: eventName}).then((data) => {
@@ -19,19 +19,20 @@ const createEvent = async (eventName, eventDate, eventLocation, eventVolunteerPr
     });
 };
 
-router.post('/create', async (req, res) => {
-    const eventName = req.body.name
-    const date = req.body.date
-    const location = req.body.location
-    const volunteerPreferences = req.body.volunteerPreferences
-    await createEvent(eventName, date, location, volunteerPreferences)
-    res.send("success")
+router.post('/create', authEndpoint.auth, async (req, res) => {
+    if(req.user && req.user.admin) {
+        const eventName = req.body.name
+        const date = req.body.date
+        const location = req.body.location
+        const volunteerPreferences = req.body.volunteerPreferences
+        await createEvent(eventName, date, location, volunteerPreferences)
+        res.sendStatus(200)
+    } else {
+        res.sendStatus(403)
+    }
  });
 
  const deleteEvent = async (eventName) => {
-
-
-
     Event.findOne({name: eventName}).then((data) => {
         if(data)
         {
@@ -45,15 +46,24 @@ router.post('/create', async (req, res) => {
     });
 };
 
- router.post('/deleteEvent', async (req, res) => {
-    const eventName = req.body.name
-    await deleteEvent(eventName);
-    res.send("success delete")
+ router.post('/deleteEvent', authEndpoint.auth, async (req, res) => { 
+    if(req.user && req.user.admin) {
+        const eventName = req.body.name
+        await deleteEvent(eventName);
+        res.sendStatus(200)
+    } else {
+        res.sendStatus(403)
+    }
  });
  
- router.get('/get-all', async(req, res) =>{
-    Event.find({}).then(events => {
-        res.json(events)
-    })
+ router.get('/get-all', authEndpoint.auth, async(req, res) =>{
+    if(req.user) {
+        Event.find({}).then(events => {
+            res.status(200)
+            res.json(events)
+        })
+    } else {
+        res.sendStatus(403)
+    }
 })
  module.exports = router;
